% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOT_calc_height.R
\name{DOT_calc_height}
\alias{DOT_calc_height}
\title{DOT_calc_height}
\usage{
DOT_calc_height(width, correction = 1.25, wh)
}
\arguments{
\item{width}{Custom width.}

\item{correction}{Pixel correction to get the exact pixel size.}

\item{wh}{List with two elements, "width" and "height".}
}
\description{
Calculate the height of the graph.
}
\author{
Sebastian Nickel
}
