% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOT_width_height.R
\name{DOT_width_height}
\alias{DOT_width_height}
\title{DOT_width_height}
\usage{
DOT_width_height(grV_obj)
}
\arguments{
\item{grV_obj}{XML of a DOT graph.}
}
\description{
Get the width and height of the graph.
}
\author{
Sebastian Nickel
}
