% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_element_from_list.R
\name{get_element_from_list}
\alias{get_element_from_list}
\title{get_element_from_list}
\usage{
get_element_from_list(list, el)
}
\arguments{
\item{list}{A list.}

\item{el}{Which element from the list should be returned}
}
\description{
Get specific element from a list.
}
\author{
Sebastian Nickel
}
