% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDOTtoPNG.R
\name{writeDOTtoPNG}
\alias{writeDOTtoPNG}
\title{writeDOTtoPNG}
\usage{
writeDOTtoPNG(graphvizobj, path, width = NULL)
}
\arguments{
\item{graphvizobj}{A DOT graph.}

\item{path}{File path to PNG file.}

\item{width}{Optional custom width. Height is set relationally.}
}
\description{
Write DOT graph to PNG
}
\author{
Sebastian Nickel
}
